requirements.txt
beautifulsoup4
bleach
certifi
chardet
Django
django-cogwheels
django-modelcluster
django-taggit
django-treebeard
djangorestframework
draftjs-exporter
html5lib
idna
lxml
Markdown
Pillow
pip
pytz
requests
setuptools
six
sqlparse
Unidecode
urllib3
wagtail
wagtail-markdown
wagtaildraftail
wagtailmenus
webencodings
Willow
uwsgi